ライブラリの処理と外部連携の一部機能を切り出した
